refactor(console): remove old top-apis ajs component
