+++ gravitee-apim-console-webui/src/management/settings/settings-routing.module.ts
@@ -29,7 +29,7 @@ import { ClientRegistrationProvidersComponent } from './client-registration-prov
 import { ClientRegistrationProviderComponent } from './client-registration-providers/client-registration-provider/client-registration-provider.component';
 import { EnvironmentMetadataComponent } from './metadata/environment-metadata.component';
 import { PortalThemeComponent } from './portal-theme/portalTheme.component';
-import { TopApisComponent as TopApisComponentMigrated } from './top-apis/migrated/top-apis.component';
+import { TopApisComponent as TopApisComponentMigrated } from './top-apis/top-apis.component';
 import { ApiLoggingComponent } from './api-logging/api-logging.component';
 import { DictionariesComponent } from './dictionaries/dictionaries.component';
 import { DictionaryComponent } from './dictionaries/dictionary.component';
