--- gravitee-apim-console-webui/src/management/settings/top-apis/dialog/add.top-api.dialog.controller.ts
@@ -1,37 +1,0 @@
-/*
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { filter, includes, map } from 'lodash';
-function AddTopApiDialogController($mdDialog: angular.material.IDialogService, ApiService, TopApiService, NotificationService, topApis) {
-  ApiService.list().then((response) => {
-    this.apis = filter(response.data, (api: any) => {
-      return !includes(map(topApis, 'api'), api.id);
-    });
-  });
-
-  this.addApi = function () {
-    TopApiService.create(this.selectedApi).then((response) => {
-      NotificationService.show("Api '" + this.selectedApi.name + "' added with success to the list of top APIs");
-      $mdDialog.hide(response.data);
-    });
-  };
-
-  this.hide = function () {
-    $mdDialog.cancel();
-  };
-}
-AddTopApiDialogController.$inject = ['$mdDialog', 'ApiService', 'TopApiService', 'NotificationService', 'topApis'];
-
-export default AddTopApiDialogController;
