--- gravitee-apim-console-webui/src/management/settings/top-apis/dialog/add.top-api.dialog.html
@@ -1,54 +1,0 @@
-<!--
-
-    Copyright (C) 2015 The Gravitee team (http://gravitee.io)
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-            http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<md-dialog aria-label="Add top API" flex="50">
-  <md-dialog-content layout-padding>
-    <h4>Add top API</h4>
-    <form>
-      <md-autocomplete
-        md-selected-item="$ctrl.selectedApi"
-        md-search-text="$ctrl.searchText"
-        md-items="api in $ctrl.apis | filter:$ctrl.searchText"
-        md-item-text="api.name"
-        placeholder="Search APIs ..."
-        md-autofocus="true"
-      >
-        <md-item-template>
-          <md-list-item class="md-2-line" style="height: 50px; min-height: 50px">
-            <gravitee-identity-picture
-              image-id="api.id"
-              image-url="api.picture_url"
-              image-name="api.name + ' ' + api.version"
-              image-width="40"
-              class="md-avatar"
-            ></gravitee-identity-picture>
-            <div class="md-list-item-text">
-              <h3>{{api.name}}</h3>
-              <p>{{item.description}}</p>
-            </div>
-          </md-list-item>
-        </md-item-template>
-        <md-not-found> No APIs matching "{{$ctrl.searchText}}" were found. </md-not-found>
-      </md-autocomplete>
-    </form>
-  </md-dialog-content>
-
-  <md-dialog-actions layout="row">
-    <md-button ng-click="$ctrl.hide()"> Cancel </md-button>
-    <md-button type="submit" class="md-raised md-primary" ng-click="$ctrl.addApi()"> Add </md-button>
-  </md-dialog-actions>
-</md-dialog>
