--- gravitee-apim-console-webui/src/management/settings/top-apis/migrated/top-apis.component.scss
@@ -1,55 +1,0 @@
-/*
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@use '../../../../scss/gio-layout' as gio-layout;
-
-:host {
-  @include gio-layout.gio-responsive-margin-container;
-  display: flex;
-  flex-direction: column;
-}
-
-.actions__edit {
-  display: flex;
-  justify-content: end;
-}
-
-.order-buttons {
-  display: flex;
-}
-
-.header {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 24px;
-
-  &__actions {
-    align-self: center;
-  }
-}
-
-.add-button {
-  width: auto;
-}
-
-.picture-cell {
-  padding: 8px 0 0 16px;
-}
-
-.avatar {
-  width: 40px;
-  height: 40px;
-}
