+++ gravitee-apim-console-webui/src/management/settings/top-apis/top-apis.component.html
