+++ gravitee-apim-console-webui/src/management/settings/top-apis/top-apis.model.ts
